Disable gcc arch optimization
By default, ImageMagick tries to optimize its build for the build host's
CPU and passes -mcpu to the compiler. This cannot be used in Fedora
since we need to support other CPUs as well, not just whatever happens
to be used on the build host. This commit disables it by passing
--without-gcc-arch to configure.

https://bugzilla.redhat.com/show_bug.cgi?id=1213828